refactor: ensure array is not empty before invoking delete actions in pinecone.service.ts
- Changed the condition to check if 'qualifyingDocs' is an empty array rather than a falsey value
- Added condition to ensure 'webVectorIds' is not empty before invoking delete actions

+++ apps/cold-platform-openai/src/pinecone/pinecone.service.ts
@@ -166,7 +166,7 @@ export class PineconeService extends BaseWorker implements OnModuleInit {
     // Filter out the matches that have a score lower than the minimum score
     const qualifyingDocs: any = matches.filter(m => m.score && m.score > minScore);
 
-    if (!qualifyingDocs) {
+    if (Array.isArray(qualifyingDocs) && qualifyingDocs.length < 1) {
       this.logger.info('No qualifying docs found', {
         message,
         docs: qualifyingDocs,
@@ -519,15 +519,17 @@ export class PineconeService extends BaseWorker implements OnModuleInit {
     const details = await this.getIndexDetails(org.name);
     const index = await this.getIndex(details.indexName);
     // delete from pinecone
-    await index.namespace(org.name).deleteMany(webVectorIds);
-    // delete from db
-    await this.prisma.vector_records.deleteMany({
-      where: {
-        id: {
-          in: webVectorIds,
+    if (webVectorIds.length > 0) {
+      await index.namespace(org.name).deleteMany(webVectorIds);
+      // delete from db
+      await this.prisma.vector_records.deleteMany({
+        where: {
+          id: {
+            in: webVectorIds,
+          },
         },
-      },
-    });
+      });
+    }
   }
 
   getCacheKey(filePayload: any) {
